// src/app/models/userinfo.ts
export class UserLoginModel {
	phonenumber: string;
	name: string;
	surname: string;
	city: string;
	address: string;
	dateofbirth: string;
	email: string;
	profilepicture: string;
	country: string;
	app_points: number;
	status_progess: string;
	last_status: string;
	tipstername: string;
	influncername: string;
	current_status: string;
	followers_count: string;
	progonisticnumber: string;
	performanceindicator: string;
	globalranking: string;
	tokenbalance: string;
	classification: string;
}

export class progonistics {
	id: string;
	name: string;
	team_awaylogo: string;
	team_homelogo: string;
	start_time: string;
	team_awayname: string;
	team_homename: string;
	starting_time: string;
	end_time: string;

}


export class Packages {
	id: string;
	name: string;
	team_awaylogo: string;
	team_homelogo: string;
	start_time: string;
	team_awayname: string;
	team_homename: string;
	starting_time: string;
	end_time: string;

}